Broke down into separate folders and files
