+++ controllers/home/home.go
@@ -1,0 +1,11 @@
+package home
+
+import (
+	"fmt"
+	"net/http"
+)
+
+func HomePage(w http.ResponseWriter, r *http.Request) {
+	fmt.Fprintf(w, "Welcome to the home page!")
+	fmt.Println("Endpoint hit: homePage")
+}
